screw it, we can split vin.py and vin_query.py later. this is nicer
